+++ argv.config.js
@@ -1,0 +1,11 @@
+var argvOptions = [
+  {
+    name: 'raw',
+    short: 'r',
+    type: 'string',
+    description: 'default output format is snazzy, use this to cancel format.',
+    example: "'script --raw' or 'script -r'",
+  }
+];
+
+module.exports = argvOptions;